refactor: make new store for open welcome
