+++ apps/designer/pages/index.tsx
@@ -1,5 +1,6 @@
 import ReactCanvas from "../components/elements/SceneComponent";
-import useStore from "../globalStore/navigationStore";
+import navigationUseStore from "../globalStore/Navigation-Store/navigationStore";
+import useStore from "../globalStore/Navigation-Store/navigationStore";
 
 // import Canvas from "../components/elements/Demo";
 // import Slider from "../components/elements/Slider";
@@ -10,7 +11,7 @@ import useStore from "../globalStore/navigationStore";
 //Canvas component is still under progress and so, has been commented out. Work Under Progres. It will contain the newer code with the functionable Zod fetching later
 
 export default function Web(): JSX.Element {
-  const { isWelcomePanelActive } = useStore();
+  const { isWelcomePanelActive } = navigationUseStore();
   return (
     <div>
       {/* <Demo/> */}
